requirements/base.txt
Django==1.8.2
django-crispy-forms==1.4.0
django_compressor==1.5
pusher==1.1.3
websockets==2.4
